+++ tests/imputers/test_ModeImputer.py
@@ -17,7 +17,10 @@ from tests.imputers.test_BaseImputer import (
     GenericImputerTransformTests,
     GenericImputerTransformTestsWeight,
 )
-from tests.utils import dataframe_init_dispatch
+from tests.utils import (
+    _convert_to_lazy,
+    dataframe_init_dispatch,
+)
 from tubular.imputers import ModeImputer
 
 
@@ -47,7 +50,8 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         "library",
         ["pandas", "polars"],
     )
-    def test_learnt_values(self, library):
+    @pytest.mark.parametrize("lazy", [True, False])
+    def test_learnt_values(self, library, lazy):
         """Test that the impute values learnt during fit are expected."""
 
         df_dict = {
@@ -60,6 +64,8 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
             "bool_col": [True, True, None, False, True],
             # cat
             "cat_col": ["b", "b", "d", "e", None],
+            # null majority
+            "null_col": [None, None, None, "a", "b"],
         }
 
         df = dataframe_init_dispatch(df_dict, library=library)
@@ -68,15 +74,18 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         df = nw.from_native(df)
         df = nw.to_native(df.with_columns(nw.col("cat_col").cast(nw.Categorical)))
 
-        x = ModeImputer(columns=["float_col", "str_col", "cat_col", "bool_col"])
+        x = ModeImputer(
+            columns=["float_col", "str_col", "cat_col", "bool_col", "null_col"]
+        )
 
-        x.fit(df)
+        x.fit(_convert_to_lazy(df, lazy=lazy))
 
         expected_impute_values = {
             "float_col": 1.0,
             "str_col": "b",
             "cat_col": "b",
             "bool_col": True,
+            "null_col": "b",
         }
 
         assert x.impute_values_ == expected_impute_values, (
@@ -87,7 +96,8 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         "library",
         ["pandas", "polars"],
     )
-    def test_learnt_values_tied(self, library):
+    @pytest.mark.parametrize("lazy", [True, False])
+    def test_learnt_values_tied(self, library, lazy):
         """Test that the impute values learnt during fit are expected - when mode is tied."""
 
         df_dict = {
@@ -111,7 +121,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         columns = ["float_col", "str_col", "cat_col", "bool_col"]
         x = ModeImputer(columns=columns)
 
-        x.fit(df)
+        x.fit(_convert_to_lazy(df, lazy=lazy))
 
         expected_impute_values = {
             "float_col": 2.0,
@@ -123,7 +133,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         with pytest.warns(
             UserWarning,  # noqa: PT030
         ) as warnings:
-            x.fit(df)
+            x.fit(_convert_to_lazy(df, lazy=lazy))
 
         columns.sort()
         warnings = [w.message.args[0] for w in warnings]
@@ -139,6 +149,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         "library",
         ["pandas", "polars"],
     )
+    @pytest.mark.parametrize("lazy", [True, False])
     @pytest.mark.parametrize(
         ("input_col", "weight_col", "learnt_value", "categorical"),
         [
@@ -159,6 +170,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         weight_col,
         learnt_value,
         categorical,
+        lazy,
     ):
         """
         Test that the impute values learnt during fit are expected -
@@ -179,14 +191,14 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         columns = ["col"]
         x = ModeImputer(columns=columns, weights_column="weight")
 
-        x.fit(df)
+        x.fit(_convert_to_lazy(df, lazy=lazy))
 
         expected_impute_values = {
             "col": learnt_value,
         }
         msg = f"ModeImputer: The Mode of column {columns[0]} is tied, will sort in descending order and return first candidate"
         with pytest.warns(UserWarning, match=msg) as warnings:
-            x.fit(df)
+            x.fit(_convert_to_lazy(df, lazy=lazy))
 
         columns.sort()
         warnings = [w.message.args[0] for w in warnings]
@@ -202,6 +214,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         "library",
         ["pandas", "polars"],
     )
+    @pytest.mark.parametrize("lazy", [True, False])
     @pytest.mark.parametrize(
         ("input_col", "weight_col", "learnt_value", "categorical"),
         [
@@ -215,6 +228,10 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
             (["a", "b", "c", "c", None], [1, 2, 3, 4, 5], "c", True),
             # float with invalid weight rows
             ([1.0, 2.0, 2.0, np.nan, 3.0, 5.0], [2, 2, 2, 1, None, -1], 2.0, False),
+            # null majority
+            ([None, None, "a", "b"], [3, 3, 2, 1], "a", False),
+            # null majority cat
+            ([None, None, "a", "b"], [3, 3, 2, 1], "a", True),
         ],
     )
     def test_learnt_values_weighted_df(
@@ -224,6 +241,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         weight_col,
         learnt_value,
         categorical,
+        lazy,
     ):
         """Test that the impute values learnt during fit are expected when df is weighted."""
         df_dict = {
@@ -241,7 +259,7 @@ class TestFit(WeightColumnFitMixinTests, GenericFitTests, FailedFitWeightFilterT
         columns = ["col"]
         x = ModeImputer(columns=columns, weights_column="weight")
 
-        x.fit(df)
+        x.fit(_convert_to_lazy(df, lazy=lazy))
 
         expected_impute_values = {
             "col": learnt_value,